CPM-64;Refactor;Renames TreePanel.tsx to tree-panel.tsx
